backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# OpenAI
openai==1.58.1

# Google Sheets
pygsheets==2.0.6
google-api-python-client==2.157.0
google-auth==2.37.0

# Environment
python-dotenv==1.0.1

# Data validation
pydantic==2.10.4

# Data processing (required by pygsheets)
pandas==2.2.3

# HTTP client (for future use)
httpx==0.28.1

# Database (Phase 5)
sqlalchemy[asyncio]==2.0.45
aiosqlite==0.22.1

# OAuth & JWT (Phase 5)
google-auth-oauthlib==1.2.3
python-jose[cryptography]==3.5.0
cryptography==46.0.3
